+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## Changelog for Cldr Utils version 2.5.0
+
+This is the changelog for Cldr Utils v2.5.0 released on October 22nd, 2019.  For older changelogs please consult the release tag on [GitHub](https://github.com/elixir-cldr/cldr_utils/tags)
+
+### Enhancements
+
+* Add `Cldr.Macros.warn_once/3` to log a warning, but only once for a given key
+
+## Changelog for Cldr Utils version 2.4.0
+
+This is the changelog for Cldr Utils v2.4.0 released on August 23rd, 2019.  For older changelogs please consult the release tag on [GitHub](https://github.com/elixir-cldr/cldr_utils/tags)
+
+### Enhancements
+
+* Add `Cldr.String.hash/1` to implement a polynomial rolling hash function
+
 ## Changelog for Cldr Utils version 2.3.0
 
 This is the changelog for Cldr Utils v2.3.0 released on June 15th, 2019.  For older changelogs please consult the release tag on [GitHub](https://github.com/elixir-cldr/cldr_utils/tags)